backend/requirements.txt
# MalGuard Backend API Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# File handling
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Security
python-jose[cryptography]>=3.3.0

# Optional: YARA support
yara-python>=4.3.0
